feat(GreatestBatteryPowerUsedFirst):battery used at the switch has the greatest amount of power
The gameplay currently selects the first battery item in the inventory to place on the switch, but since dead batteries can be collected, there needs to be a way where the batteries that are not dead are used first. The solution to this is to sort the inventory items in descending power order so that the player doesn't use a dead battery to power a switch when there are other functioning batteries in the player's inventory.