feat(interactableNotes):Notes can be viewed during game
When the player is close enough to a note, the player is able to press E and view the note on a separate panel. The panel displays the content and the player is able to scroll through it if it doesn't all fit on the screen. While the player is reading the note, the game is paused so the sanity and power count of the game is frozen until the player closes the panel.