feat(abstractNoteController):notecontroller is abstract
The NoteController script can now be attached to a game object and once the appropriate gameobjects and images are dragged into it, it will display the image to the player when the player presses the E key. You also need to make sure the collider attached to the interactable game object is big enough so that the player can trigger the call.